% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authentication.R
\name{Auth_OpenAI}
\alias{Auth_OpenAI}
\title{Set up authentication using OpenAI API key}
\usage{
Auth_OpenAI(key, organization = NULL, return_list = FALSE)
}
\arguments{
\item{key}{Your OpenAI API key.}

\item{organization}{Usage from the API requests will count against the specified organization's subscription quota. Set to null as default.}

\item{return_list}{Whether return the OpenAI available model list?}
}
\description{
Input the OpenAI API key and the account organization (optimal) to set up the authentication.
Just ignore the \code{return_list} parameter, it is used to export the model list available from OpenAI.
}
