% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_func_char.R
\docType{data}
\name{chat_func_char}
\alias{chat_func_char}
\title{ChatGPT function written in string for dynamic parsing.}
\format{
An object of class \code{character} of length 1.
}
\usage{
chat_func_char
}
\description{
ChatGPT function written in string for dynamic parsing.
}
\keyword{datasets}
