% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_session_class.R
\docType{class}
\name{chat_session-class}
\alias{chat_session-class}
\title{chat_session class}
\description{
chat_session class is used for prompting messages to ChatGPT model and archiving chat history.
}
