% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authentication.R
\name{OpenAI_model_list}
\alias{OpenAI_model_list}
\title{Get OpenAI available model list.}
\usage{
OpenAI_model_list(simplify = TRUE)
}
\arguments{
\item{simplify}{Whether simplify the list returned?}
}
\value{
A character vector or a complex character list.
}
\description{
Get OpenAI available model list.
}
