% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_URL.R
\docType{data}
\name{URL_chat}
\alias{URL_chat}
\title{OpenAI chat API URL.}
\format{
An object of class \code{character} of length 1.
}
\usage{
URL_chat
}
\description{
OpenAI chat API URL.
}
\keyword{datasets}
