% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_URL.R
\docType{data}
\name{URL_completion}
\alias{URL_completion}
\title{OpenAI completion API URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
URL_completion
}
\description{
OpenAI completion API URL.
}
\keyword{datasets}
