% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_URL.R
\docType{data}
\name{URL_models}
\alias{URL_models}
\title{OpenAI all models URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
URL_models
}
\description{
OpenAI all models URL.
}
\keyword{datasets}
